requirements.txt
click
rich
requests
types-requests
setproctitle
pydantic==1.10.2
pyyaml
types-PyYAML
